requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
huggingface-hub==0.26.5
requests==2.32.3
pydantic==2.10.2
httpx==0.27.2
aiohttp==3.11.8
SQLAlchemy==2.0.36
transformers==4.48.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
TTS==0.22.0
pydub==0.25.1
noisereduce==3.0.3
librosa==0.10.0
soundfile==0.13.0
google-genai==0.8.0
elevenlabs==1.50.6
boto3==1.36.2